Install dockerd rootless without supplied scripts
On a host system, dockerd rootless is installed with the
dockerd-rootless-setuptool.sh script. When using in dind mode the
setuptool script no longer works because rootlesskit is run in order to
test that the install was successful. This isn't possible because
rootlesskit would need to run privileged at build time.

The rootlesskit supplied install scripts are rewritten to only include
whats needed to run in dind mode.