Map all system groups into rootless user namespace
Instead of trying to specifically map the docker group we map the entire
space the system groups. The way subgid mapping works is that it
sequential maps a range of groups on the host into the user namespace.
For example if `rootless:998:1` is the first entry in subgid then the
host group 998 gets mapped to user namespace group 1. Mapping all the
system groups ensures the order stays the same as the host and any
system group docker is using it will get mapped